Add Setting for Local Plugin Path
This adds a setting (that can only be set in `kana.json`) called
localPluginsPath that if `type` is set to `theme` will map the directory
specified to the plugins directory, allowing you to set the plugins you
want/need installed for development of the theme.

It also adds the _stub_ of a setting called `localPlugins` that will
be used like the `plugins` setting but to install the plugin(s) in the path
specified in the `localPluginsPath` but the setting does nothing currently.